@@ -10,9 +10,9 @@
     "auth0": "^0.8.2",
     "auth0-common-logging": "auth0/auth0-common-logging#v2.11.0",
     "auth0-styleguide": "auth0/styleguide#2.2.7",
-    "auth0-tutorial-navigator": "auth0/tutorial-navigator#1.1.0",
+    "auth0-tutorial-navigator": "auth0/tutorial-navigator#2.0.1",
     "aws-kinesis-writable": "^1.3.1",
-    "babel": "^5.8.35",
+    "babel": "^5.8.38",
     "body-parser": "^1.15.0",
     "bunyan": "^1.5.1",
     "cb": "^0.1.0",
@@ -22,20 +22,21 @@
     "express": "^4.13.4",
     "express-redirect": "^1.2.2",
     "express-session": "^1.13.0",
-    "fluxible": "^1.0.4",
-    "fluxible-addons-react": "^0.2.5",
+    "fluxible": "^1.1.0",
+    "fluxible-addons-react": "^0.2.8",
+    "fluxible-plugin-devtools": "^0.1.3",
     "fluxible-plugin-service-proxy": "^1.1.0",
-    "fluxible-router": "^0.4.8",
+    "fluxible-router": "^0.4.14",
     "gray-matter": "^2.0.1",
     "helmet": "^1.3.0",
-    "highlight.js": "^9.2.0",
+    "highlight.js": "^9.3.0",
     "httpsnippet": "^1.16.5",
     "jade": "^1.11.0",
-    "js-yaml": "^3.5.3",
+    "js-yaml": "^3.6.0",
     "limitd-client": "^1.8.1",
-    "lodash": "^4.5.1",
+    "lodash": "^4.11.1",
     "lsr": "^1.0.0",
-    "markdown-it": "^6.0.0",
+    "markdown-it": "^6.0.1",
     "markdown-it-container": "^2.0.0",
     "method-override": "^2.3.5",
     "mongo-getdb": "^2.0.0",
@@ -46,14 +47,14 @@
     "passport": "^0.2.2",
     "protobufjs": "^4.0.0",
     "randomstring": "^1.0.7",
-    "react": "^0.14.7",
-    "react-dom": "^0.14.7",
-    "request": "^2.69.0",
+    "react": "^15.0.1",
+    "react-dom": "^15.0.1",
+    "request": "^2.72.0",
     "retry": "^0.8.0",
-    "serialize-javascript": "^1.1.2",
+    "serialize-javascript": "^1.2.0",
     "server-destroy": "^1.0.1",
-    "sitemap": "^1.5.0",
-    "url-join": "0.0.1",
+    "sitemap": "^1.6.0",
+    "url-join": "^1.1.0",
     "web-header": "auth0/web-header#0.5.4",
     "whatwg-fetch": "^0.11.0",
     "winston": "^1.0.1",
@@ -83,7 +84,7 @@
     "babel-polyfill": "^6.1.4",
     "bower": "^1.4.1",
     "bundle-loader": "^0.5.4",
-    "clean-webpack-plugin": "^0.1.4",
+    "clean-webpack-plugin": "^0.1.9",
     "css-loader": "^0.23.1",
     "es6-promise": "^3.0.2",
     "exports-loader": "^0.6.2",
@@ -102,9 +103,10 @@
     "progress": "^1.1.8",
     "react-hot-loader": "^1.3.0",
     "style-loader": "^0.13.0",
-    "stylus-loader": "^1.4.0",
+    "stylus": "^0.54.5",
+    "stylus-loader": "^2.0.0",
     "url-loader": "^0.5.6",
-    "webpack": "^1.12.2",
+    "webpack": "^1.13.0",
     "webpack-dev-server": "^1.12.1",
     "yaml-loader": "^0.1.0"
   }
