@@ -2,6 +2,7 @@ import getWidgetScriptUrl from './widget-script-url';
 import nconf from 'nconf';
 import getDb from '../../data';
 import express from 'express';
+import winston from 'winston';
 
 var router = express.Router();
 
@@ -21,7 +22,10 @@ DEMOS.forEach(function (demo) {
     };
 
     getWidgetScriptUrl(params, function (err, assetsUrl, tenant_domain, namespace, client) {
-      if (err) return res.send(err);
+      if (err && process.env.NODE_ENV !== 'production') {
+        // This happens on local development
+        return res.send('This frame does not render for local development. This is normal.');
+      }
 
       client = client || {
         clientID: WIDGET_FALLBACK_CLIENTID,
